fixes to delete, working on put request

+++ index.js
@@ -69,13 +69,27 @@ server.post('/api/users', (req, res) => {
 
 server.delete('/api/users/:id', (req, res)=>{
     const { id } = req.params;
-    db.remove(id)
+    const user = req.body;
+
+    if(!user){
+        res.status(404).json({message: "The user with the specified ID does not exist."})
+    } else{
+        db.remove(id)
         .then(removedUser => {
             res.json(removedUser)
         })
         .catch(err => {
             res.status(500).json({error: "The user could not be removed."})
         })
+    }
+})
+
+
+
+// Put request
+
+server.put('/api/users/:id', (req, res) => {
+
 })
 
 // Listening
